Time: 34 ms (81.71%), Space: 106 MB (13.02%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.cpp
@@ -5,14 +5,14 @@ public:
         long long left = 0;
         long long right = 0;
         long long sum = 0;
-        long long ans = 1;
+        int ans = 1;
         while(right < nums.size()) {
             sum += nums[right];
             while (nums[right] * (right - left + 1) - sum > k) {
                 sum -= nums[left];
                 left++;
             }
-            ans = max(ans, right - left + 1);
+            ans = max(ans, (int)(right - left + 1));
             right++;
         }
         return ans;
